move network to network-old
